TRUNK-4971 Fixes and refactoring
